Revert "Merge branch 'dev' of git://github.com/Baystation12/Baystation12 into Baystation12-dev"
This reverts commit 82806aa161b42a9f974dd7d59d8c2c6a4a9f6b40.

+++ code/game/objects/items/devices/radio/radio.dm
@@ -207,6 +207,8 @@
 	var/datum/radio_frequency/connection = null
 	if(channel && channels && channels.len > 0)
 		if (channel == "department")
+//			log_debug(channel=\"[channel]\" switching to \"[channels[1]]\"")
+
 			channel = channels[1]
 		connection = secure_radio_connections[channel]
 	else
@@ -214,10 +216,14 @@
 		channel = null
 	if (!istype(connection))
 		return
+	if (!connection)
+		return
+
 	var/mob/living/silicon/ai/A = new /mob/living/silicon/ai(src, null, null, 1)
 	A.fully_replace_character_name(from)
 	talk_into(A, message, channel,"states")
 	qdel(A)
+	return
 
 // Interprets the message mode when talking into a radio, possibly returning a connection datum
 /obj/item/device/radio/proc/handle_message_mode(mob/living/M as mob, message, message_mode)
@@ -267,6 +273,8 @@
 	//#### Grab the connection datum ####//
 	var/datum/radio_frequency/connection = handle_message_mode(M, message, channel)
 	if (!istype(connection))
+		return 0
+	if (!connection)
 		return 0
 
 	var/turf/position = get_turf(src)
@@ -321,6 +329,7 @@
 
 
   /* ###### Radio headsets can only broadcast through subspace ###### */
+
 	if(subspace_transmission)
 		// First, we want to generate a new radio signal
 		var/datum/signal/signal = new
@@ -429,6 +438,7 @@
 
 	//THIS IS TEMPORARY. YEAH RIGHT
 	if(!connection)	return 0	//~Carn
+
 	return Broadcast_Message(connection, M, voicemask, pick(M.speak_emote),
 					  src, message, displayname, jobname, real_name, M.voice_name,
 					  filter_type, signal.data["compression"], list(position.z), connection.frequency,verb,speaking)
@@ -711,7 +721,7 @@
 	invisibility = 101
 	listening = 0
 	canhear_range = 0
-	channels=list("Engineering" = 1, "Security" = 1, "Medical" = 1, "Command" = 1, "Common" = 1, "Science" = 1, "Supply" = 1, "Service" = 1)
+	channels=list("Engineering","Security", "Medical", "Command")
 
 /obj/item/device/radio/announcer/Destroy()
 	crash_with("attempt to delete a [src.type] detected, and prevented.")
